Crate netfuse

netfuse - FUSE abstractions for a network filesystem

Implementing a network-backed filesystem requries implementing netfuse::NetworkFilesystem and calling netfuse::mount with your implementation.

Internally, the call to netfuse::mount will create a low-level NetFuse filesystem that handles caching, inode number to path translation, offsets and sizes, and lazily triggering writes to persist.

Structs

DirEntry
Entry from a directory listing

Metadata
Metadata representing a file

MountOptions
Options for configuring how the NetworkFilesystem will be mounted

NetFuse
Low-level FUSE implementation that is backed by an implementation of NetworkFilesystem

Traits

NetworkFilesystem
Trait to implement to provide a backend store for a NetFuse filesystem

Functions

mount
Mount the given NetworkFilesystem. This function will not return until the filesystem is unmounted.

Type Definitions

LibcError
libc Error Code