Struct netfuse::NetFuse

pub struct NetFuse<NFS: NetworkFilesystem> {
    // some fields omitted
}

Low-level FUSE implementation that is backed by an implementation of NetworkFilesystem

The NetFuse implementation manages the the inode store, including a mapping between inode number and path. It also provides a data cache, and the abstraction that manages read/write offsets and lengths, as well as lazy persistence.